Refactor utils to a subpackage.
